+++ src/app/components/BackgroundSection.jsx
@@ -57,8 +57,8 @@ const BackgroundSection = () => {
                 src={EDUCATION.logo}
                 alt="Chapman Logo"
                 className="text-sm font-semibold self-start mt-auto"
-                width={isMd ? 150 : 150}
-                height={isMd ? 75 : 75}
+                width={isMd ? 150 : 75}
+                height={isMd ? 150 : 75}
               />
             </div>
             <div>
