+++ src/app/utils/displayDate.js
@@ -4,8 +4,9 @@ import { monthToNumber } from "./constants";
 export const displayDate = (startMonth, startYear, endMonth, endYear) => {
   // Empty string if end date is before start date
   if (
-    endYear < startYear ||
-    (monthToNumber[endMonth] < monthToNumber[endYear] && endYear <= startYear)
+    parseInt(endYear) < parseInt(startYear) ||
+    (monthToNumber[endMonth] < monthToNumber[startMonth] &&
+      parseInt(endYear) <= parseInt(startYear))
   )
     return "";
