+++ src/app/utils/duration.js
@@ -12,7 +12,7 @@ export const calculateDuration = (startMonth, stateYear, endMonth, endYear) => {
   const monthsDiff = endDate.getMonth() - startDate.getMonth();
 
   // Edge case: end date starts before start date
-  if (yearsDiff < 0 || (monthsDiff < 0 && yearsDiff < 0))
+  if (yearsDiff < 0 || (monthsDiff < 0 && yearsDiff <= 0))
     return { years: 0, months: 0 };
 
   // Months are inclusive
