+++ src/tests/utils/displayDate.test.js
@@ -6,7 +6,9 @@ describe("displayDate", () => {
   });
 
   it("returns empty string if end date is before start date", () => {
-    expect(displayDate("Jan", "2023", "Jan", "2022")).toEqual("");
+    expect(displayDate("Jan", "2023", "Jan", "2022")).toEqual(""); // Year
+    expect(displayDate("Feb", "2023", "Jan", "2023")).toEqual(""); // Month
+    expect(displayDate("Feb", "2023", "Jan", "2022")).toEqual(""); // Month and year
   });
 
   it("returns formatted string when endMonth and endYear are not provided (Present)", () => {
